Move ddsi shm functions to shm_transport.h
