+++ src/core/ddsi/src/q_init.c
@@ -72,7 +72,7 @@
 #ifdef DDS_HAS_SHM
 #include "dds/ddsrt/io.h"
 #include "iceoryx_binding_c/runtime.h"
-#include "dds/ddsi/shm_init.h"
+#include "dds/ddsi/shm_transport.h"
 #endif
 
 static void add_peer_addresses (const struct ddsi_domaingv *gv, struct addrset *as, const struct ddsi_config_peer_listelem *list)
